refactor: split select(bottom sheet / base) (close #182)
